requirements.txt
matplotlib
numpy
pytorch